move app state action env reducer to AppCore
